+++ lib/collab-client/src/collab-extension.ts
@@ -5,10 +5,10 @@ import { Plugin } from '@bangle.dev/pm';
 import { uuid } from '@bangle.dev/utils';
 
 import { collabClientPlugin } from './collab-client';
-import { hardResetClient, queryFatalError } from './commands';
+import { hardResetClient, queryCollabState, queryFatalError } from './commands';
 
 export const plugins = pluginsFactory;
-export const commands = { queryFatalError, hardResetClient };
+export const commands = { queryFatalError, hardResetClient, queryCollabState };
 
 export interface CollabExtensionOptions {
   requestTimeout?: number;