+++ lib/collab-client/src/index.ts
@@ -1,3 +1,7 @@
 import * as collabClient from './collab-extension';
 
+export type { CollabStateName, FatalErrorCode } from './common';
+
 export { collabClient };
+
+export type { ValidCollabState } from './state';